finished todoApp for firecast firebase

+++ functions/APIs/todos.js
@@ -4,6 +4,7 @@ const { db } = require('../util/admin');
 
 exports.getAllTodos = (request, response) => {
   db.collection('todos')
+    .where('username', '==', request.user.username)
     .orderBy('createdAt', 'desc')
     .get()
     .then((data) => {
@@ -12,6 +13,7 @@ exports.getAllTodos = (request, response) => {
         todos.push({
           todoId: doc.id,
           title: doc.data().title,
+          username: doc.data().username,
           body: doc.data().body,
           createdAt: doc.data().createdAt,
         });
@@ -21,6 +23,25 @@ exports.getAllTodos = (request, response) => {
     .catch((err) => {
       console.error(err);
       return response.status(500).json({ error: err.code });
+    });
+};
+
+exports.getOneTodo = (request, response) => {
+  db.doc(`/todos/${request.params.todoId}`)
+    .get()
+    .then((doc) => {
+      if (!doc.exists) {
+        return response.status(404).json({
+          error: 'Todo not found',
+        });
+      }
+      TodoData = doc.data();
+      TodoData.todoId = doc.id;
+      return response.json(TodoData);
+    })
+    .catch((err) => {
+      console.error(err);
+      return response.status(500).json({ error: error.code });
     });
 };
 
@@ -37,6 +58,7 @@ exports.postOneTodo = (request, response) => {
     title: request.body.title,
     body: request.body.body,
     createdAt: new Date().toISOString(),
+    username: request.user.username,
   };
 
   db.collection('todos')
@@ -60,6 +82,11 @@ exports.deleteTodo = (request, response) => {
       if (!doc.exists) {
         return response.status(404).json({ error: 'Todo not found' });
       }
+
+      if (doc.data().username !== request.user.username) {
+        return response.status(403).json({ error: 'UnAuthorized' });
+      }
+
       return document.delete();
     })
     .then(() => {
